Proposed slightly different solution
Hey, thanks for your article and example code. After working on it and testing (right now on iOS only, but real device) I've observed that if dealing with images, works better waiting for document to be loaded entirely.
I also made more changes, because I'm loading a remote string html-shaped, but wanted to share this one with you. For example, adding a width limit to images on injected CSS, depending on width of the device.

Thanks again, very smart solution!